# Add online KNN, SVD and ALS