--- a/mangaki/mangaki/tests/test_rating.py
+++ b/mangaki/mangaki/tests/test_rating.py
@@ -1,6 +1,7 @@
 from django.test import TestCase
 from django.core.urlresolvers import reverse
 from django.contrib.auth import get_user_model
+
 from mangaki.models import Studio, Category, Rating, Editor, Work
 
 
@@ -23,3 +24,12 @@ class RatingTest(TestCase):
         self.assertEqual(Rating.objects.get(user__username='test', work_id=self.anime.id).choice, 'favorite')
         self.client.post(vote_url, {'choice': 'favorite'})  # Cancel my vote
         self.assertFalse(Rating.objects.filter(user__username='test', work_id=self.anime.id).exists())
+
+    def test_anonymized_rating(self):
+        vote_url = reverse('vote', args=[self.anime.id])
+        response = self.client.post(vote_url, {'choice': 'like'})
+        self.assertEqual(response.content, b'like')
+
+        profile_url = reverse('my-profile', args=['anime', 'seen'])
+        response = self.client.get(profile_url)
+        self.assertIn(self.anime.title, response.content.decode('utf-8'))