--- a/mangaki/mangaki/utils/fit_algo.py
+++ b/mangaki/mangaki/utils/fit_algo.py
@@ -1,6 +1,8 @@
+from django.conf import settings
+# from mangaki.models import Rating
+
 from zero.dataset import Dataset
 from zero.recommendation_algorithm import RecommendationAlgorithm
-from django.conf import settings
 
 
 def fit_algo(algo_name, triplets, titles=None, categories=None,